Make interpreter analyze a set of sources in one go instead of sequentially.
The interpreter supports passing set of sources using --load.
When one starts to spead definitions over packages, it would be very annoying if
one has to topologically sort packages in dependency order. What is more, if
two packages mutually recursive dependencies, then they have to be analyzed together.

This commit changes the organization of source in the interpreter slightly to
permit loading of the multiple --load sources in one go.

PiperOrigin-RevId: 491843123