Fix type inference/checking of relation types.
We previously used "approximate" type checking, which can find many
errors but is unsound. A simple declaration with bounds
[/number /string] [/string /number]
would be turned into a single relation type like
[fn:Union(/number, /string), fn:Union(/number, /string)]
and accept an argument tuple /number /number.
This CL does away with such approximation and adds the proper representation
and handling of alternatives. In short, given a set of alternatives for
a relation, we try all these possibilities when assigning types to the
variables that appear in a clause. If we run out of possibilities, we report
a type error.

PiperOrigin-RevId: 502860920