Add :match_prefix, "reflects" descriptor, rewriting & typing.
A unary predicate can be declared to "reflect" a name prefix type.
It has to be ensured that the arguments are name constants such that
1. match the name prefix type and 2. there are no name constants
that match the name prefix which do not also appear as arguments.
In other words, the predicate arguments is exactly the set of name
constants that can match the name prefix type. For this reason
the predicate is said to "reflect" the type.

We also introduce a new builtin predicate :match_prefix which can
play a similar role. Under some conditions, premises of a rule
can be rewritten so that look-up operations / filters get replaced
with a simple prefix check.

Also adds a few missing headers.

PiperOrigin-RevId: 514987786