Type checking: add function type, add ctx parameter.
This is still work in progress. The FunType is not yet used anywhere.
We will really have two (two and a half) categories of type expression:

- the "set expressions" with base types, structured types, unions. The "rel types" are merely tuples of set expression used to check predicates foo(X, Y). These describe the values that a mangle variable can take in the body of a rule.

- the "(functional) type expression" with type variables, base types, structured types, function types. We want a way to call builtin functions (which is no longer datalog) and later external functions from the host language. For type-checking such expressions, we need a function type.

Logic programming rules interact with the functional expressions in very
limited and predictable ways (still, giving up datalog safety in most case).
When a function is called, we should always be able to infer the exact
argument types.

PiperOrigin-RevId: 521747363