Optimize memory usage of multi indexed fact store
The multi indexed fact store stores the Atoms in a map indexed by the parameter index and value. If an Atom has multiple parameters, then it is stored multiple times. The parameters are usually ast.Contant Go structs, that store the value inline. This means the data is copied multiple times. Change the map to store only the pointer, so the original data is not duplicated.
This decreases the memory usage by 50% if most atoms have 2 parameters.

PiperOrigin-RevId: 539104703