Fix: neg. atoms must be delayed until all variables are bound.
It is convenient in Mangle code to switch between the view that
rules are declarative (order does not matter, what the user expects)
and explicitly ordered (what the implementation does). In this
case, this switch led to a bug where something that is ok
in the declarative view (negative atom that relies on variables being
bound later) was not ok in the explicit ordered view. The fix
is to simply reorder when such a situation is detected.

PiperOrigin-RevId: 591210513