Support for deferred predicates and non-powerset lattice. Fixes #24. Removing will happen if the FactStore is an instance of FactStoreWithRemove, otherwise Remove will be a no-op.
See map_lattice_test.go for an example how to use.
In order to use merging, a fundep and a merge predicate has to
be specified in the declaration. The fundep declares a functional
dependency X1,...,Xn --> Y1,...,Ym and the merge predicate
is a (possibly deferred) three-place predicate Yin1, Yin2, Yout.
that merges (applies the lattice-join) to Yin1 and Yin2, yielding
the result in Yout.
Currently, only functional dependencies with a single target column
are supported (m=1). This is not a real limitation as the column
could be a tuple.

PiperOrigin-RevId: 594247353