Add type inference for :match_field
While we can give a polymorphic relation type for built-in
relations like :match_cons at some point and treat those
uniformly, doing the same for `:match_field` would involve
a dependent type "given a record type A and field f,
return field type of A.f".

So we add special treatment :match_field operates on
a scrutinee that must have been bound before, so we
extract the required field type when we can.

PiperOrigin-RevId: 653641773