Grammar change add better syntax for  types, especially struct types. Fixes #27
".Type<args>" translates to 'fn:Type(args)'. This is the existing (weird)
syntax, so this change only affects parsing and readability.

Not that the ".Struct<members>" syntax supports "term : term" as well as
"opt term : term" definitions. These get mapped to "term, term" (two
arguments) and "fn:opt(term, term)" (one argument).

This change only adapts the go implementation.
The Rust implementation is left for later.

We could think of types offering the full (positional arguments), (named arguments), variadic and keyword variadic that is known from python and starlark. This
flexibility will only be interesting if there is also a way to declare new
generic types.

PiperOrigin-RevId: 717809486