In aggregations, permit calls of normal (non-reducer) functions.
We check that argument variables have a single value.
This is the case when they appear either in the group-by
or they are defined by previous statements.

One could actually permit the grouped to be list-valued.
This is left for later.

This addresses a TODO and makes rules a lot more succinct.
In terms of monad operations, this corresponds to "map".

PiperOrigin-RevId: 805808461