Add support for external predicates, fixes #59
ExternalPredicateCallback is used to query external data sources.

An atom `mydb(input1, ..., inputN, OutputVar1, ..., OutputVarN)` is evaluated
as follows:
- the engine checks whether the fact store contains any facts
  of the shape `mydb(input1, ..., inputN, _, ..., _)`. If so, we
  use those for evaluation.
- if no facts were found, the engine calls `ShouldQuery(input1, ..., inputN)`
  if false, evaluation continues.
- if true, the engine calls `Query(input1, ..., inputN, filter1, ..., filterM)`
  and expects outputs callback. Every output tuple gets added
  as `mydb(input1, ..., inputN, output1, ..., outputM)` fact to the store,
  and continues evaluation.
  if ExecuteQuery returns an error, evaluation fails with that error.
If tuples (input1, ..., inputN) are known to yield empty results, the
implementation can keep track of that and prevent an unnecessary
call to `ExecuteQuery`.
filters contains the arguments that are output positions which are either
variables or constants that are used to match the position (filters proper).
The implementation may use the constant filter arguments for filter-pushdown,
but is also free to ignore them. In any case, when constant filter arguments
are present, only matching facts will be added to the store.

PiperOrigin-RevId: 810639975